requirements.txt
flask
newsapi-python==0.2.6
gunicorn
googletrans==3.1.0a0